refactor portfolio workspace route ownership
